Extract interface for immediate TaskSpec running
