+++ async/src/main/java/org/odk/collect/async/CoroutineAndWorkManagerScheduler.kt
@@ -27,30 +27,15 @@ class CoroutineAndWorkManagerScheduler(
         workManager
     ) // Needed for Java construction
 
+    private val taskSpecRunner = WorkManagerTaskSpecImmediateRunner(workManager)
+
     override fun immediate(
         tag: String,
         spec: TaskSpec,
         inputData: Map<String, String>,
         notificationInfo: NotificationInfo
     ) {
-        val workManagerInputData = Data.Builder()
-            .putString(TaskSpecWorker.DATA_TASK_SPEC_CLASS, spec.javaClass.name)
-            .putBoolean(TaskSpecWorker.FOREGROUND, true)
-            .putString(TaskSpecWorker.FOREGROUND_NOTIFICATION_CHANNEL, notificationInfo.channel)
-            .putString(
-                TaskSpecWorker.FOREGROUND_NOTIFICATION_CHANNEL_NAME,
-                notificationInfo.channelName
-            )
-            .putInt(TaskSpecWorker.FOREGROUND_NOTIFICATION_TITLE, notificationInfo.title)
-            .putInt(TaskSpecWorker.FOREGROUND_NOTIFICATION_ID, notificationInfo.id)
-            .putAll(inputData)
-            .build()
-
-        val workRequest = OneTimeWorkRequest.Builder(TaskSpecWorker::class.java)
-            .addTag(tag)
-            .setInputData(workManagerInputData)
-            .build()
-        workManager.beginUniqueWork(tag, ExistingWorkPolicy.REPLACE, workRequest).enqueue()
+        taskSpecRunner.run(tag, spec, inputData, notificationInfo)
     }
 
     override fun networkDeferred(
