+++ async/src/main/java/org/odk/collect/async/TaskSpecImmediateRunner.kt
@@ -1,0 +1,5 @@
+package org.odk.collect.async
+
+interface TaskSpecImmediateRunner {
+    fun run(tag: String, taskSpec: TaskSpec, inputData: Map<String, String>, notificationInfo: NotificationInfo)
+}
